# docs/api/openapi-mvp3-draft.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ontology Platform Backend",
    "description": "Ontology Platform MVP backend API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/validation-jobs": {
      "post": {
        "tags": [
          "Validation"
        ],
        "summary": "Create validation job",
        "operationId": "create_validation_job_api_v1_projects__project_id__validation_jobs_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationJobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationJob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      },
      "get": {
        "tags": [
          "Validation"
        ],
        "summary": "List validation jobs",
        "operationId": "list_validation_jobs_api_v1_projects__project_id__validation_jobs_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationJob"
                  },
                  "title": "Response List Validation Jobs Api V1 Projects  Project Id  Validation Jobs Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation-jobs/{validation_job_id}": {
      "get": {
        "tags": [
          "Validation"
        ],
        "summary": "Get validation job",
        "operationId": "get_validation_job_api_v1_validation_jobs__validation_job_id__get",
        "parameters": [
          {
            "name": "validation_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Validation Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationJob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validation-jobs/{validation_job_id}/results": {
      "get": {
        "tags": [
          "Validation"
        ],
        "summary": "List validation results",
        "operationId": "list_validation_results_api_v1_validation_jobs__validation_job_id__results_get",
        "parameters": [
          {
            "name": "validation_job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Validation Job Id"
            }
          },
          {
            "name": "candidate_kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CandidateKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ValidationResult"
                  },
                  "title": "Response List Validation Results Api V1 Validation Jobs  Validation Job Id  Results Get"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "CandidateKind": {
        "type": "string",
        "enum": [
          "ENTITY",
          "RELATION",
          "PROPERTY_VALUE"
        ],
        "title": "CandidateKind"
      },
      "CandidateRef": {
        "properties": {
          "candidate_kind": {
            "$ref": "#/components/schemas/CandidateKind"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          }
        },
        "type": "object",
        "required": [
          "candidate_kind",
          "candidate_id"
        ],
        "title": "CandidateRef"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationJob": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "ontology_version_id": {
            "type": "string",
            "title": "Ontology Version Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "extraction_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/ValidationJobStatus"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "summary": {
            "$ref": "#/components/schemas/ValidationSummary"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "ontology_version_id",
          "status",
          "created_at",
          "summary"
        ],
        "title": "ValidationJob"
      },
      "ValidationJobCreateRequest": {
        "properties": {
          "ontology_version_id": {
            "type": "string",
            "title": "Ontology Version Id"
          },
          "extraction_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extraction Job Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "candidate_refs": {
            "items": {
              "$ref": "#/components/schemas/CandidateRef"
            },
            "type": "array",
            "title": "Candidate Refs"
          },
          "rule_codes": {
            "items": {
              "$ref": "#/components/schemas/ValidationRuleCode"
            },
            "type": "array",
            "title": "Rule Codes"
          }
        },
        "type": "object",
        "required": [
          "ontology_version_id"
        ],
        "title": "ValidationJobCreateRequest"
      },
      "ValidationJobStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCESS",
          "FAILED"
        ],
        "title": "ValidationJobStatus"
      },
      "ValidationResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "validation_job_id": {
            "type": "string",
            "title": "Validation Job Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "ontology_version_id": {
            "type": "string",
            "title": "Ontology Version Id"
          },
          "candidate_kind": {
            "$ref": "#/components/schemas/CandidateKind"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "rule_code": {
            "$ref": "#/components/schemas/ValidationRuleCode"
          },
          "severity": {
            "$ref": "#/components/schemas/ValidationResultSeverity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field_path": {
            "type": "string",
            "title": "Field Path"
          },
          "blocking": {
            "type": "boolean",
            "title": "Blocking"
          },
          "suggested_fix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Fix"
          },
          "details": {
            "type": "object",
            "title": "Details"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "validation_job_id",
          "project_id",
          "ontology_version_id",
          "candidate_kind",
          "candidate_id",
          "rule_code",
          "severity",
          "message",
          "field_path",
          "blocking",
          "created_at"
        ],
        "title": "ValidationResult",
        "example": {
          "blocking": false,
          "candidate_id": "candidate-entity-demo",
          "candidate_kind": "ENTITY",
          "created_at": "2026-06-19T00:00:00Z",
          "details": {
            "confidence": 0.6
          },
          "field_path": "confidence",
          "id": "validation-result-demo",
          "message": "Candidate confidence is below review threshold.",
          "ontology_version_id": "ontology-version-demo",
          "project_id": "project-demo",
          "rule_code": "LOW_CONFIDENCE",
          "severity": "WARNING",
          "suggested_fix": "Confirm the candidate against evidence.",
          "validation_job_id": "validation-job-demo"
        }
      },
      "ValidationResultSeverity": {
        "type": "string",
        "enum": [
          "INFO",
          "WARNING",
          "FAILED"
        ],
        "title": "ValidationResultSeverity"
      },
      "ValidationRuleCode": {
        "type": "string",
        "enum": [
          "CLASS_EXISTS",
          "RELATION_EXISTS",
          "RELATION_DOMAIN_RANGE",
          "RELATION_DIRECTION",
          "REQUIRED_PROPERTY",
          "DATATYPE",
          "CARDINALITY",
          "DUPLICATE_CANDIDATE",
          "ORPHAN_NODE",
          "EVIDENCE_MISSING",
          "ONTOLOGY_VERSION_MISMATCH",
          "LOW_CONFIDENCE"
        ],
        "title": "ValidationRuleCode"
      },
...
      "ValidationSummary": {
        "properties": {
          "target_count": {
            "type": "integer",
            "title": "Target Count",
            "default": 0
          },
          "passed_count": {
            "type": "integer",
            "title": "Passed Count",
            "default": 0
          },
          "warning_count": {
            "type": "integer",
            "title": "Warning Count",
            "default": 0
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count",
            "default": 0
          },
          "missing_evidence_count": {
            "type": "integer",
            "title": "Missing Evidence Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "ValidationSummary"
      },
...
    }
  }
}